% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{WBG Secretariat - Style fonts
Tables and charts -
Titles: General Sans Bold
Subtitles: Switzer Bold or Regular
Spanners, stubs or groupings (or column 1) and captions: Satoshi
Numbers and figures: Open Sans}
\usage{
load_fonts(
  general_sans_regular = file.path(here::here("Fonts", "GeneralSans-Regular.otf")),
  general_sans_italic = file.path(here::here("Fonts", "GeneralSans-Italic.otf")),
  general_sans_bold = file.path(here::here("Fonts", "GeneralSans-Bold.otf")),
  general_sans_bolditalic = file.path(here::here("Fonts", "GeneralSans-BoldItalic.otf")),
  switzer_regular = file.path(here::here("Fonts", "Switzer-Regular.otf")),
  switzer_bold = file.path(here::here("Fonts", "Switzer-Bold.otf")),
  switzer_italic = file.path(here::here("Fonts", "Switzer-Italic.otf")),
  switzer_bolditalic = file.path(here::here("Fonts", "Switzer-BoldItalic.otf")),
  satoshi_regular = file.path(here::here("Fonts", "Satoshi-Regular.otf")),
  satoshi_bold = file.path(here::here("Fonts", "Satoshi-Bold.otf")),
  satoshi_italic = file.path(here::here("Fonts", "Satoshi-Italic.otf")),
  open_sans_regular = file.path(here::here("Fonts", "OpenSans-Regular.ttf")),
  open_sans_italic = file.path(here::here("Fonts", "OpenSans-Italic.ttf"))
)
}
\arguments{
\item{general_sans_regular}{path to General Sans font (regular)}

\item{general_sans_italic}{path to General Sans font (italic)}

\item{general_sans_bold}{path to General Sans font (bold)}

\item{general_sans_bolditalic}{path to General Sans font (bolditalic)}

\item{switzer_regular}{path to Switzer font (regular)}

\item{switzer_bold}{path to Switzer  font (bold)}

\item{switzer_italic}{path to Switzer font (italic)}

\item{switzer_bolditalic}{path to Switzer font (bolditalic)}

\item{satoshi_regular}{path to Satoshi font (regular)}

\item{satoshi_bold}{path to Satoshi font (bold)}

\item{satoshi_italic}{path to Satoshi font (italic)}

\item{open_sans_regular}{path to Open Sans font (regular)}

\item{open_sans_italic}{path to Open Sans font (italic)}
}
\description{
WBG Secretariat - Style fonts
Tables and charts -
Titles: General Sans Bold
Subtitles: Switzer Bold or Regular
Spanners, stubs or groupings (or column 1) and captions: Satoshi
Numbers and figures: Open Sans
}
