% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{sec_color}
\alias{sec_color}
\title{HEX code of a color in the SEC color palette}
\usage{
sec_color(which_color)
}
\arguments{
\item{which_color}{The name of the color in the palette}
}
\value{
The hex code of each of the colors in the palette
}
\description{
\code{sec_color} returns the hex code(s) that correspond(s) to
the name(s) of the color(s) in the SEC palette that are
present in its arguments.
}
\details{
Either a single character string or a vector of strings
can be passed through, with the function returning an equal
number of hex codes.
}
