% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablethemes.R
\name{sec_tabtheme}
\alias{sec_tabtheme}
\title{WBG SEC Table Themes - a GT theme with the color schemes
and styles of the World Bank Group's Corporate Secretariat}
\usage{
sec_tabtheme(gt_object, ...)
}
\arguments{
\item{gt_object}{The table to which the theme will be applied}

\item{...}{Special to pass undetermined number of arguments}
}
\value{
A table styled using the WB SEC theme
}
\description{
WBG SEC Table Themes - a GT theme with the color schemes
and styles of the World Bank Group's Corporate Secretariat
}
\examples{
\dontrun{

library(tidyverse)
library(gt)

# Classic example using mtcars, but applying formatting to show
# the available tweaks to titles for best control

 iris |>
 group_by(Species) |>
 summarise(mean(Sepal.Length)) |>
 gt() |>
 tab_header(title = html(
 "<span style = 'background-color: #00AFEF; color: white'>Figure 1</span>
   Iris Dataset"),
           subtitle = html(
 "Lorem Ipsum is simply dummy text that is typicallly used in the <br>
  &nbsp&nbspand typesetting <i>industry</i>.<br><br>")) |>
 tab_source_note(source_note = md(
 "**Source:** Corporate Secretariat | **Data:** mtcars")) |>

 sec_tabtheme()

}
}
