% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthemes.R
\name{theme_wbgsec}
\alias{theme_wbgsec}
\title{A \link{ggplot} theme with a specific}
\usage{
theme_wbgsec(
  axes = "",
  grid = "",
  legend_position = "right",
  caption_position = "right",
  title_family = "General Sans",
  subtitle_family = "Switzer",
  axis_title_family = "Satoshi",
  axis_text_family = "Satoshi",
  legend_title_family = "Satoshi",
  legend_text_family = "Satoshi",
  facet_title_family = "Switzer",
  caption_family = "General Sans",
  title_size = 18,
  subtitle_size = 11,
  axis_title_size = 10,
  axis_text_size = 9,
  legend_title_size = 10,
  legend_text_size = 10,
  facet_title_size = 10,
  caption_size = 9,
  title_color = "#00AFEF",
  subtitle_color = "#00218F",
  axis_title_color = "#404040",
  axis_text_color = "#404040",
  legend_title_color = "#404040",
  legend_text_color = "#404040",
  facet_title_color = "#404040",
  caption_color = "#404040",
  background_color = "#FFFFFF",
  axis_line_color = "#a6a6a6",
  grid_color = "#dad5d1"
)
}
\arguments{
\item{axes}{A string indicating which axes should have lines and ticks.
Specify which axes to show by including the matching characters in the
string: "t" for top, "r" for right, "b" for bottom, "l" for left. You will
need to ensure this argument is consistent with the axes settings in your
plot for the lines and ticks to be displayed. The default is "", meaning
no ticks no lines for the vertical or horizontal axes are shown by default.}

\item{grid}{A string indicating which gridlines should be shown. Specify
the gridlines to show by including the matching characters in the string:
"h" for horizontal, "v" for vertical. The default is an empty string,
meaning no gridlines are shown by default.}

\item{legend_position}{A string indicating the position of the legend. Valid
positions are "top", "right", "bottom", "left", "top-right", "top-left",
"bottom-right", "bottom-left", and "none". The default is "right".}

\item{caption_position}{A string indicating the horizontal position of the
caption. Valid positions are "left" or "right". The default is "right".}

\item{title_family}{A string indicating the font-family to use for the
title. The default is "National-LFSN Semibd".}

\item{subtitle_family}{A string indicating the font-family to use for the
subtitle. The default is "National-LFSN Book".}

\item{axis_title_family}{A string indicating the font-family to use for
axis titles. The default is "Open Sans".}

\item{axis_text_family}{A string indicating the font-family to use for
axis text. The default is "Open Sans".}

\item{legend_title_family}{A string indicating the font-family to use for
legend titles. The default is "Open Sans".}

\item{legend_text_family}{A string indicating the font-family to use for
legend text. The default is "Open Sans".}

\item{facet_title_family}{A string indicating the font-family to use for
facet titles. The default is "Open Sans".}

\item{caption_family}{A string indicating the font-family to use for
captions. The default is "National-LFSN Book".}

\item{title_size}{An integer indicating the font size to use for the title
in points. The default is 18 points.}

\item{subtitle_size}{An integer indicating the font size to use for the
subtitle in points. The default is 14 points.}

\item{axis_title_size}{An integer indicating the font size to use for axis
titles in points. The default is 11 points.}

\item{axis_text_size}{An integer indicating the font size to use for axis
text in points. The default is 10 points.}

\item{legend_title_size}{An integer indicating the font size to use for
legend titles in points. The default is 10 points.}

\item{legend_text_size}{An integer indicating the font size to use for
legend text in points. The default is 10 points.}

\item{facet_title_size}{An integer indicating the font size to use for
facet titles in points. The default is 10 points.}

\item{caption_size}{An integer indicating the font size to use for captions
in points. The default is 12 points.}

\item{title_color}{An RGB hex string indicating the color to use for the
title. The default is "#006548".}

\item{subtitle_color}{An RGB hex string indicating the color to use for the
subtitle. The default is "#006548".}

\item{axis_title_color}{An RGB hex string indicating the color to use for
axis titles. The default is "#404040".}

\item{axis_text_color}{An RGB hex string indicating the color to use for
axis text. The default is "#404040".}

\item{legend_title_color}{An RGB hex string indicating the color to use for
legend titles. The default is "#404040".}

\item{legend_text_color}{An RGB hex string indicating the color to use for
legend text. The default is "#404040".}

\item{facet_title_color}{An RGB hex string indicating the color to use for
facet titles. The default is "#404040".}

\item{caption_color}{An RGB hex string indicating the color to use for
captions. The default is "#404040".}

\item{background_color}{An RGB hex string indicating the color to use for
the background. The default is "#f0eeed".}

\item{axis_line_color}{An RGB hex string indicating the color to use for
the axis lines. The default is "#a6a6a6".}

\item{grid_color}{An RGB hex string indicating the color to use for the
gridlines. The default is "#dad5d1".}
}
\value{
A ggplot2 theme that implements the Commons Library style.
}
\description{
A \link{ggplot} theme with a specific
}
\examples{
\dontrun{

library(tidyverse)
iris |>
ggplot(aes(x = Species, y = Sepal.Length)) +
geom_point(color = sec_color("secpink")) +
labs(title = "Figure 1. Iris Data Set",
     subtitle = "Lorem ipsum is simply dummy text that is used in the typesetting industry",
     x = "Species of Iris",
     y = "Length of the Sepals") +
theme_wbgsec() +
}

}
